requirements.txt
# requirements.txt
pydantic
PyYAML
apscheduler
requests
Pillow
pytz
